fix: prevent negative source index from int truncation
Widen source_index to ptrdiff_t and drop the (int) casts so large
partition indices no longer truncate to negative values, and guard
source insertion with >= 0 instead of the != -1 sentinel.

+++ include/coordinator.h
@@ -25,7 +25,7 @@ typedef struct {
   size_t start_coords[DIMENSIONS];
   size_t problem_sizes[DIMENSIONS];
   unsigned int iterations;
-  int source_index;
+  ptrdiff_t source_index;
   size_t absorption_size;
 } dc_partition_info_t;
